@@ -10,15 +10,20 @@
   },
   "dependencies": {
     "@apollo/client": "^3.7.14",
+    "@reduxjs/toolkit": "^2.2.1",
     "autoprefixer": "10.4.14",
     "bootstrap-icons": "^1.10.4",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
+    "eslint-config-next": "^14.1.2",
     "graphql": "^16.6.0",
-    "next": "13.3.0",
+    "next": "^14.1.2",
     "postcss": "8.4.21",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
+    "react-toastify": "^10.0.4",
+    "redux": "^5.0.1",
+    "redux-persist": "^6.0.0",
     "tailwindcss": "3.3.1"
   }
 }
