@@ -3,64 +3,80 @@ export const menuItems = [
       id: 1,
       imageUrl: "/media/images/products/spaghetti.png",
       name: "Spaghetti garnished with fruit toppings",
+      slug: "spaghetti-garnished-with-fruit-toppings",
       excerpt: "A classic Nigerian dish made with long grain rice, tomatoes, onions, and a blend of spices. Served with your choice of protein",
       price: 1999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 2,
       imageUrl: "/media/images/products/chicken-suya.png",
       name: "Grilled flavoured chicken suya",
+      slug: "grilled-flavoured-chicken-suya",
       excerpt: "A hearty soup made with ground melon seeds, vegetables, and your choice of protein. Served with fufu or rice.",
       price: 1799.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 3,
       imageUrl: "/media/images/products/vegetable-salad.webp",
       name: "Vegetable Salad",
+      slug: "vegetable-salad",
       excerpt: "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Tempore quos, nemo atque facilis odit officiis.",
       price: 2999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 4,
       imageUrl: "/media/images/products/shredded-peppered-meat.png",
       name: "Shredded Peppered Meat",
+      slug: "shredded-peppered-meat",
       excerpt: "A popular street food made from black-eyed peas, onions, and spices, deep-fried until crispy on the outside and soft on the inside. Served with a side of spicy pepper sauce",
       price: 2999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 5,
       imageUrl: "/media/images/products/assorted-peppered-chicken.png",
       name: "Assorted Peppered Chicken",
+      slug: "assorted-peppered-chicken",
       excerpt: "A flavorful soup made with your choice of meat or fish, simmered with spices, herbs, and vegetables. Served hot and spicy",
       price: 2999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 6,
       imageUrl: "/media/images/products/bread-and-akara.png",
       name: "Bread and Akara",
+      slug: "bread-akara",
       excerpt: "A classic Nigerian dish made with yam pounded into a smooth, dough-like consistency,",
       price: 2999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 7,
       imageUrl: "/media/images/products/noodles-with-fish-and-egg.png",
       name: "Noodles with fish and Eggs",
+      slug: "noodles-with-fish-eggs",
       excerpt: "A classic Nigerian dish made with yam pounded into a smooth, dough-like consistency.",
       price: 2999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
     {
       id: 8,
       imageUrl: "/media/images/products/assorted-kpomo.webp",
       name: "Peppered Kpomo",
+      slug: "peppered-kpomo",
       excerpt: "A delicious and nutritious vegetable stew made with spinach, bell peppers, onions, and your choice of protein. Served with rice or fufu.",
       price: 2999.99,
-      featured: true
+      featured: true,
+      availableQuantity: 5
     },
 ];
